rename project to cocktail guru
