+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Viper.iml" filepath="$PROJECT_DIR$/Viper.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/cocktailguru.iml" filepath="$PROJECT_DIR$/cocktailguru.iml" />
     </modules>
   </component>
 </project>
