+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.dylange.viper">
+          package="com.dylange.cocktailguru">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".main.MainActivity"
+            android:name="com.dylange.cocktailguru.main.MainActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -19,13 +19,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".base.BaseActivity"
+            android:name="com.dylange.cocktailguru.base.BaseActivity"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".splash.SplashActivity"
+            android:name="com.dylange.cocktailguru.splash.SplashActivity"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".login.LoginActivity"
+            android:name="com.dylange.cocktailguru.login.LoginActivity"
             android:screenOrientation="portrait"/>
     </application>
