+++ app/src/main/java/com/dylange/cocktailguru/login/LoginActivity.kt
@@ -1,12 +1,12 @@
-package com.dylange.viper.login
+package com.dylange.cocktailguru.login
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import com.dylange.viper.R
-import com.dylange.viper.base.BaseActivity
-import com.dylange.viper.data.DataManager
-import com.dylange.viper.main.MainActivity
+import com.dylange.cocktailguru.R
+import com.dylange.cocktailguru.base.BaseActivity
+import com.dylange.cocktailguru.data.DataManager
+import com.dylange.cocktailguru.main.MainActivity
 
 /**
  * Created by dylanlange on 27/10/17.
