+++ app/src/main/java/com/dylange/cocktailguru/login/LoginPresenter.kt
@@ -1,4 +1,4 @@
-package com.dylange.viper.login
+package com.dylange.cocktailguru.login
 
 import android.os.Bundle
 
@@ -9,7 +9,7 @@ import android.os.Bundle
 class LoginPresenter(override val mView: LoginContract.View,
                      override val mInteractor: LoginContract.Interactor): LoginContract.Presenter {
 
-	override fun onCreateCalled(savedInstanceState: Bundle?, extras: Bundle) {
+	override fun onCreateCalled(savedInstanceState: Bundle?, extras: Bundle?) {
 
 	}
