+++ app/src/main/java/com/dylange/cocktailguru/main/MainContract.kt
@@ -1,6 +1,6 @@
-package com.dylange.viper.main
+package com.dylange.cocktailguru.main
 
-import com.dylange.viper.base.BaseContract
+import com.dylange.cocktailguru.base.BaseContract
 
 /**
  * Created by dylanlange on 27/10/17.
