+++ app/src/main/java/com/dylange/cocktailguru/splash/SplashInteractor.kt
@@ -1,6 +1,6 @@
-package com.dylange.viper.splash
+package com.dylange.cocktailguru.splash
 
-import com.dylange.viper.data.DataManagerContract
+import com.dylange.cocktailguru.data.DataManagerContract
 
 /**
  * Created by dylanlange on 27/10/17.
