+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.dylange.viper.main.MainActivity">
+    tools:context="com.dylange.cocktailguru.main.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
