+++ app/src/test/java/com/dylange/cocktailguru/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.dylange.viper
+package com.dylange.cocktailguru
 
 import org.junit.Test
